Move Enums to tag_db.py
