+++ src/tribler/core/components/tag/restapi/tests/test_tags_endpoint.py
@@ -7,7 +7,8 @@ from ipv8.keyvault.crypto import default_eccrypto
 from pony.orm import db_session
 
 from tribler.core.components.restapi.rest.base_api_test import do_request
-from tribler.core.components.tag.community.tag_payload import TagOperation, TagOperationEnum, TagRelationEnum
+from tribler.core.components.tag.community.tag_payload import TagOperation
+from tribler.core.components.tag.db.tag_db import TagOperationEnum, TagRelationEnum
 from tribler.core.components.tag.restapi.tags_endpoint import TagsEndpoint
 from tribler.core.conftest import TEST_PERSONAL_KEY
 from tribler.core.utilities.unicode import hexlify
